+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to
 
 
 ## [Unreleased]
+### Added
+- support for setting file attributes on files, symbolic links and directories
+  in `populate-file-tree`.
 
 ## [0.1.8] — 2019-08-26
 ### Added